Reverse gzip extension matching logic
We now attempt to gzip everything apart from explicitly excluded
extensions.

Given that the command already checks whether compression has resulted
in any saving this should be a safer default. The exclusion list is now
just an optimisation so we don't bother trying to compress files where
we know we won't make a saving.