Handle invalid byte sequences in URLs
Previously these would trigger an error, now they are ignored.

This is hard to test for as depending on the WSGI server you are using
such sequences are often filtered out before WhiteNoise ever sees them,
however I have reproduced the issue using gunicorn and confirmed that
the following patch fixes the problem.

Closes #112