Log names of compressed files during collectstatic
This is done by yielding the names up to the calling management command
which handles the actual logging for us.

Now that compression of all files happens together at the end of the
post-processing sequence it's more important to give the user visibility
into what's happening. Otherwise it would just look like the process was
hanging for ages when it ought to have finished.