Fix HTTP 304 handling when both ETag and last modified specified
Previously if the ETag specified in the `If-None-Match` header
differed from the file's ETag, but the last modified time in the
`If-Modified-Since` header was newer than the file modified time,
then an HTTP 304 would be returned rather than the expected HTTP 200.

This case can occur when a deployment is rolled back, and User Agents
make a request with the reverted content's ETag and newer last
modified time. (Browsers typically send both headers.)

Fixes #202.